break inspector in own package
